word size fixups
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@34468 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSZone.m
@@ -102,8 +102,7 @@
 void *
 GSOutOfMemory(NSUInteger size, BOOL retry)
 {
-  fprintf(stderr, "GSOutOfMemory ... wanting %lu bytes.\n",
-    (unsigned long)size);
+  fprintf(stderr, "GSOutOfMemory ... wanting %"PRIuPTR" bytes.\n", size);
   return 0;
 }
